requirements.txt
aioredis==1.3.1
apns2==0.7.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-rest-auth==1.1.1
Django==3.1
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-csp==3.7
django-debug-toolbar==2.2
django-extensions==3.0.4
django-extra-fields==3.0.0
django-fakery==2.3.1
django-filter==2.3.0
django-guardian==2.3.0
django-import-export==2.3.0
django-jalali==4.0.0
django-jalali-date==0.3.1
django-jazzmin==2.2.7
django-money==1.1
django-oauth-toolkit==1.3.2
django-online-users==0.3
django-pagedown==2.1.2
django-push-notifications==2.0.0
django-redis==4.12.1
django-reorder==0.2.1
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
django-simple-history==2.11.0
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.1.1
flake8==3.8.3
gunicorn==20.0.4
h2==2.6.2
hiredis==1.1.0
hpack==3.0.0
http-ece==1.1.0
humanize==2.5.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.0
install==1.3.3
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jdatetime==3.6.2
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
netwell==0.4.0
notebook==6.1.1
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
persiantools==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
public==0.0.0
py-moneyed==0.8.0
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python3-openid==3.2.0
pytz==2020.1
pywebpush==1.11.0
PyYAML==5.3.1
pyzmq==19.0.2
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
timedelta==0.0.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
